+++ history.md
@@ -1,6 +1,10 @@
 # 2022-08-28
 
 - Make `CocMenuSel` use background of `PmenuSel`.
+- Snippet related configuration changed (old configuration still works until next release)
+  - "coc.preferences.snippetStatusText" -> "snippet.statusText"
+  - "coc.preferences.snippetHighlight" -> "snippet.highlight"
+  - "coc.preferences.nextPlaceholderOnDelete" -> "snippet.nextPlaceholderOnDelete"
 
 # 2022-08-27